requirements.txt
opencv-python
<<<<<<< HEAD
opencv-python-headless
Pillow
=======
>>>>>>> e588fd2393a1409b343b36f1f9eb4a3e6f132ab9
numpy
pyperclip
